functions/proxy/requirements.txt
firebase-functions>=0.4.0
firebase-admin>=6.0.0
requests>=2.28.0
python-dotenv
google-cloud-firestore==2.13.0
google-cloud-storage==2.13.0
google-cloud-tasks==2.14.0
google-cloud-translate==3.12.1
google-generativeai==0.3.1